Check the SHA-256 of the downloaded file
Check to see if the SHA-256 of the downloaded file matches the value in
the SHA256SUMS file.  This check fails if the downloaded file does not
have a corresponding line in the SHA256SUMS file.  Unfortunately, TinyGo
does not sign releases, so this trust-on-first-use approach is the best
we have today.